api/requirements.txt
fastapi
uvicorn
websockets
pyzmq